+++ src/main/java/net/bannerretrieval/CauldronBehaviorExtended.java
@@ -15,28 +15,25 @@ import net.minecraft.util.ItemActionResult;
 public interface CauldronBehaviorExtended {
     CauldronBehavior WASH_SHIELD = (state, world, pos, player, hand, stack) -> {
         BannerPatternsComponent bannerPatternsComponent = stack.getOrDefault(DataComponentTypes.BANNER_PATTERNS, BannerPatternsComponent.DEFAULT);
-        ComponentMap bannerComponents = stack.getOrDefault(BannerRetrievalComponents.BANNER_COMPONENTS, ComponentMap.EMPTY);
         DyeColor bannerColor = stack.get(DataComponentTypes.BASE_COLOR);
         assert bannerColor != null;
 
-        if (bannerPatternsComponent.layers().isEmpty() || bannerComponents.isEmpty()) {
+        if (bannerPatternsComponent.layers().isEmpty()) {
             return ItemActionResult.PASS_TO_DEFAULT_BLOCK_INTERACTION;
         }
         if (world.isClient) return ItemActionResult.success(true);
 
         ItemStack shieldItemStack = stack.copy();
 
-        shieldItemStack.remove(BannerRetrievalComponents.BANNER_COMPONENTS);
         shieldItemStack.remove(DataComponentTypes.BASE_COLOR);
-        //shieldItemStack.remove(DataComponentTypes.BANNER_PATTERNS);
         shieldItemStack.set(DataComponentTypes.BANNER_PATTERNS, BannerPatternsComponent.DEFAULT);
 
         ItemStack bannerItemStack = new ItemStack(bannerItemFromColor(bannerColor));
 
-        bannerItemStack.applyComponentsFrom(bannerComponents);
+        bannerItemStack.set(DataComponentTypes.BANNER_PATTERNS, bannerPatternsComponent);
 
         stack.decrementUnlessCreative(1, player);
-        if (shieldItemStack.isEmpty()) {
+        if (stack.isEmpty()) {
             player.setStackInHand(hand, shieldItemStack);
         } else if (player.getInventory().insertStack(shieldItemStack)) {
             player.playerScreenHandler.syncState();