+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-# Update to 1.21
+# Buxfixes
+
+- Made the mod server-only at the cost of not retaining the original banner components.